requirements.txt
aiosqlite==0.17.0
anyio==3.3.4
asgiref==3.4.1
asyncpg==0.24.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
ecdsa==0.17.0
fastapi==0.70.0
h11==0.12.0
idna==3.3
iso8601==0.1.16
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pypika-tortoise==0.1.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
